moved AskException to ditto-base-model as the exception must be available everywhere (e.g. gateway/connectivity) * fixed AbstractEnforcementWithAsk exception handling * adjusted *UnavailableException's descriptions to perform retry to improve resiliency
